+++ lib/contentful/management/client.rb
@@ -20,9 +20,7 @@ require 'contentful/management/client_upload_methods_factory'
 require 'contentful/management/client_api_key_methods_factory'
 require 'contentful/management/client_webhook_methods_factory'
 require 'contentful/management/client_snapshot_methods_factory'
-require 'contentful/management/client_api_usage_methods_factory'
 require 'contentful/management/client_environment_methods_factory'
-require 'contentful/management/client_usage_period_methods_factory'
 require 'contentful/management/client_organization_methods_factory'
 require 'contentful/management/client_content_type_methods_factory'
 require 'contentful/management/client_ui_extension_methods_factory'
@@ -31,7 +29,9 @@ require 'contentful/management/client_webhook_health_methods_factory'
 require 'contentful/management/client_preview_api_key_methods_factory'
 require 'contentful/management/client_space_membership_methods_factory'
 require 'contentful/management/client_editor_interface_methods_factory'
+require 'contentful/management/client_space_periodic_usage_methods_factory'
 require 'contentful/management/client_personal_access_tokens_methods_factory'
+require 'contentful/management/client_organization_periodic_usage_methods_factory'
 
 module Contentful
   module Management
@@ -133,22 +133,20 @@ module Contentful
         ClientOrganizationMethodsFactory.new(self)
       end
 
-      # Allows viewing of usage periods in context of the current client
-      # Allows listing all usage periods for client.
+      # Allows listing all usage periods for organization grouped by organization.
       # @see _ README for details.
       #
-      # @return [Contentful::Management::ClientUsagePeriodMethodsFactory]
-      def usage_periods(organization_id)
-        ClientUsagePeriodMethodsFactory.new(self, organization_id)
+      # @return [Contentful::Management::ClientOrganizationPeriodicUsageMethodsFactory]
+      def organization_periodic_usages(organization_id)
+        ClientOrganizationPeriodicUsageMethodsFactory.new(self, organization_id)
       end
 
-      # Allows viewing of api usage in context of the current client
-      # Allows listing all api usage for client.
+      # Allows listing all usage periods for organization grouped by space.
       # @see _ README for details.
       #
-      # @return [Contentful::Management::ClientApiUsageMethodsFactory]
-      def api_usage(organization_id)
-        ClientApiUsageMethodsFactory.new(self, organization_id)
+      # @return [Contentful::Management::ClientSpacePeriodicUsageMethodsFactory]
+      def space_periodic_usages(organization_id)
+        ClientSpacePeriodicUsageMethodsFactory.new(self, organization_id)
       end
 
       # Allows viewing of users in context of the current client